+++ wishbone/lookup/choice.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-#  __init__.py
+#  choice.py
 #
 #  Copyright 2016 Jelle Smet <development@smetj.net>
 #
@@ -22,16 +22,32 @@
 #
 #
 
+from wishbone.lookup import Lookup
+from random import choice as choice_array
 
-class Lookup(object):
-    pass
 
-from .event import EventLookup
-from .choice import Choice
-from .cycle import Cycle
-from .etcd import ETCD
-from .pid import PID
-from .random_bool import RandomBool
-from .random_integer import RandomInteger
-from .random_word import RandomWord
-from .random_uuid import RandomUUID
+class Choice(Lookup):
+
+    '''
+    **Returns a random element from the provided array.**
+
+    This function returns a random element from the provided array.
+
+    - Parameters to initialize the function:
+
+        - values(list)(None): An array of elements to choose from
+
+    - Parameters to call the function:
+
+        None
+
+
+    '''
+
+    def __init__(self, array):
+
+        self.array = array
+
+    def lookup(self):
+
+        return choice_array(self.array)